refactor : 설문 응답 페이지 리팩토링
