+++ src/components/survey/components/TurnAPageButtons.jsx
@@ -1,12 +1,30 @@
 import React from "react";
+import { useSelector, useDispatch } from "react-redux";
 import styled from "styled-components";
 
-const TurnAPageButtons = ({
-  currentPageNum,
-  questionLength,
-  goBackPageClickHandler,
-  nextPageClickHandler,
-}) => {
+import {
+  getCover,
+  __getBeforeSurveyQuestion,
+  __getSurveyQuestion,
+} from "../../../redux/modules/surveySlice";
+
+const TurnAPageButtons = () => {
+  const dispatch = useDispatch();
+  const questionIdList = useSelector((state) => state.survey.questionIdList);
+  const currentPageNum = useSelector((state) => state.survey.currentPageNum);
+
+  const nextPageClickHandler = () => {
+    currentPageNum === questionIdList.length + 1
+      ? alert("마지막 항목입니다")
+      : dispatch(__getSurveyQuestion(questionIdList[currentPageNum - 1]));
+  };
+
+  const goBackPageClickHandler = () => {
+    currentPageNum === 2
+      ? dispatch(getCover())
+      : dispatch(__getBeforeSurveyQuestion(questionIdList[currentPageNum - 3]));
+  };
+
   return (
     <ArrowContainer>
       <img
@@ -15,7 +33,7 @@ const TurnAPageButtons = ({
         onClick={goBackPageClickHandler}
       />
       <div>
-        {currentPageNum}/{questionLength}
+        {currentPageNum}/{questionIdList.length + 1}
       </div>
       <img
         src={process.env.PUBLIC_URL + "/img/phoneRightArrow.svg"}
@@ -27,17 +45,15 @@ const TurnAPageButtons = ({
 };
 
 const ArrowContainer = styled.div`
-  width: 26.3rem;
   display: flex;
   justify-content: space-between;
   align-items: center;
+  width: 26.3rem;
   margin: 0 auto;
   img {
     &:nth-child(1) {
       width: 1.6rem;
       cursor: pointer;
-    }
-    &:nth-child(2) {
     }
     &:nth-child(3) {
       width: 1.6rem;
