+++ src/components/survey/components/typeOfSurvey/CoverSurvey.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { useDispatch } from "react-redux";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
+
 import fonts from "../../../../styles/fonts";
 import { __getSurveyQuestion } from "../../../../redux/modules/surveySlice";
 import RoundButtonLarge from "../../../common/buttons/roundButtons/RoundButtonLarge";
@@ -10,6 +10,7 @@ const CoverSurvey = () => {
   const dispatch = useDispatch();
   const survey = useSelector((state) => state.survey.survey);
   const questionIdList = useSelector((state) => state.survey?.questionIdList);
+
   const surveyStartClickHandler = () => {
     dispatch(__getSurveyQuestion(questionIdList[0]));
   };
@@ -33,40 +34,41 @@ const CoverSurvey = () => {
 };
 
 const Container = styled.div`
-  width: 100%;
-  height: 100%;
   position: relative;
   display: flex;
   flex-direction: column;
+  width: 100%;
+  height: 100%;
 `;
 
 const Main = styled.div`
   flex: 1;
-  text-align: center;
-  padding-top: 6.1rem;
   display: flex;
   flex-direction: column;
+  padding-top: 6.1rem;
   height: 20rem;
+
+  text-align: center;
   h1 {
-    ${fonts.Body1}
     margin: 0;
+    ${fonts.Body1}
     font-weight: 700;
     font-size: 2.4rem;
     line-height: 2.9rem;
   }
   h5:nth-of-type(1) {
+    margin-top: 4.6rem;
     ${fonts.Body1}
     font-weight: 500;
     font-size: 1.6rem;
     line-height: 1.9rem;
-    margin-top: 4.6rem;
   }
   h5:nth-of-type(2) {
+    margin-top: 2rem;
     ${fonts.Body3}
     font-weight: 500;
     font-size: 1.6rem;
     line-height: 1.9rem;
-    margin-top: 2rem;
   }
 `;
