+++ src/components/survey/components/typeOfSurvey/LongDescriptiveSurvey.jsx
@@ -1,21 +1,14 @@
 import React from "react";
 import { useSelector, useDispatch } from "react-redux";
 import styled from "styled-components";
+
+import { changeDescriptive } from "../../../../redux/modules/surveySlice";
 import fonts from "../../../../styles/fonts";
-import {
-  changeDescriptive,
-  getCover,
-} from "../../../../redux/modules/surveySlice";
-import {
-  __getSurveyQuestion,
-  __getBeforeSurveyQuestion,
-} from "../../../../redux/modules/surveySlice";
+import Title from "../Title";
 import TurnAPageButtons from "../../components/TurnAPageButtons";
 
 const LongDescriptiveSurvey = () => {
   const dispatch = useDispatch();
-  const question = useSelector((state) => state.survey.question);
-  const questionIdList = useSelector((state) => state.survey.questionIdList);
   const currentPageNum = useSelector((state) => state.survey.currentPageNum);
   const descriptive = useSelector(
     (state) => state.survey?.answer[currentPageNum - 2]["descriptive"]
@@ -26,24 +19,9 @@ const LongDescriptiveSurvey = () => {
     dispatch(changeDescriptive(answer));
   };
 
-  const nextPageClickHandler = () => {
-    currentPageNum === questionIdList.length + 1
-      ? alert("마지막 항목입니다")
-      : dispatch(__getSurveyQuestion(questionIdList[currentPageNum - 1]));
-  };
-
-  const goBackPageClickHandler = () => {
-    currentPageNum === 2
-      ? dispatch(getCover())
-      : dispatch(__getBeforeSurveyQuestion(questionIdList[currentPageNum - 3]));
-  };
-
   return (
     <Container>
-      <TitleContainer>
-        <h1>{question.questionTitle}</h1>
-        <h5>{question.questionSummary}</h5>
-      </TitleContainer>
+      <Title />
       <InputContainer>
         <textarea
           type="text"
@@ -53,12 +31,7 @@ const LongDescriptiveSurvey = () => {
         ></textarea>
       </InputContainer>
       <ArrowButtonContainer>
-        <TurnAPageButtons
-          currentPageNum={currentPageNum}
-          questionLength={questionIdList.length + 1}
-          goBackPageClickHandler={goBackPageClickHandler}
-          nextPageClickHandler={nextPageClickHandler}
-        />
+        <TurnAPageButtons />
       </ArrowButtonContainer>
     </Container>
   );
@@ -74,52 +47,32 @@ const Container = styled.div`
   padding-top: 6.1rem;
 `;
 
-const TitleContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  h1 {
-    ${fonts.Body1}
-    margin: 0;
-    font-weight: 700;
-    font-size: 2.4rem;
-    line-height: 2.9rem;
-  }
-  h5 {
-    ${fonts.Body3}
-    font-weight: 500;
-    font-size: 1.6rem;
-    line-height: 1.9rem;
-    margin-top: 4.6rem;
-  }
-`;
-
 const InputContainer = styled.div`
-  width: 100%;
   display: flex;
-  margin-top: 5rem;
   justify-content: center;
   align-items: center;
+
+  width: 100%;
+  margin-top: 5rem;
   textarea {
     box-sizing: border-box;
+
     width: 26.4rem;
     height: 23.7rem;
     padding: 1rem;
-    resize: none;
-
-    font-size: 18px;
-    background-color: transparent;
-    color: gray;
-    border: ${({ theme }) => `2px solid ${theme.subColor1}`};
-
-    border-radius: 1rem;
-    scroll-behavior: auto;
 
     ${fonts.Body1}
     font-weight: 600;
     font-size: 1.4rem;
     line-height: 1.7rem;
 
+    background-color: transparent;
+    color: ${({ theme }) => theme.gray2};
+    border: ${({ theme }) => `2px solid ${theme.subColor1}`};
+    border-radius: 1rem;
+
+    scroll-behavior: auto;
+    resize: none;
     &::placeholder {
       color: gray;
       text-align: center;
@@ -132,8 +85,8 @@ const InputContainer = styled.div`
 
 const ArrowButtonContainer = styled.div`
   position: absolute;
-  width: 100%;
   bottom: 5rem;
+  width: 100%;
 `;
 
 export default LongDescriptiveSurvey;
