+++ src/components/survey/screen/SurveyScreen.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+
 import SurveyView from "../components/SurveyView";
 
 const SurveyScreen = () => {
