+++ src/pages/Survey.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import styled from "styled-components";
+
 import SurveyScreen from "../components/survey/screen/SurveyScreen";
 
 const Survey = () => {
